rename: dateFormat -> date-format
